Correcciones de datos, formatos de fechas, formatos de fecha en los PDF, estado de los préstamos, cambios en las restricciones de las tablas, restricciones para borrar tanto las carreras como los roles, entre otros.

+++ app/Http/Controllers/CarreraController.php
@@ -96,6 +96,11 @@ class CarreraController extends Controller
     public function destroy($id)
     {
         $carreras = Carrera::find($id);
+
+        if ($carreras->ficha()->count() > 0) {
+            return redirect()->back()->with('mensaje', 'No se puede eliminar la carrera porque tiene fichas asociadas.')->with('icono', 'error');
+        }
+
         $carreras->delete();
 
         return redirect()->route('admin.carreras.index')
